[FIX] mail: chat bubble not cut with spreadsheet scrollbar
Before this commit, when using spreadsheet with some chat bubbles,
the chat bubbles were cut with horizontal scrollbar of spreadsheet.

Steps to reproduce:
- make a spreadsheet
- put a chat in chat bubble (e.g. from click on messaging menu)
- remove the "filters" panel of the spreadsheet

This happens because the horizontal scrollbar of spreadsheet uses
a `z-index: 15` while chat bubble uses `z-index: 10`.

The `z-index` of chat bubble is relative to navigable list, and
since the value was exactly slightly offset from navigable list,
this commit fixes by shifting the `$o-mail-NavigableList-zIndex`.
It was `11`, now it's `21`. As a result, the chat bubble `z-index`
is being changed from `10` to `20`, which is greater to `15` from
spreadsheet horizontal scrollbar.

Task-4547672

closes odoo/odoo#198021

Signed-off-by: Alexandre Kühn (aku) <aku@odoo.com>